Move atributo RegisteredArticles.registered_titles_and_authors para ArticlesMergence e remove a classe RegisteredArticles
